pyproject.toml
[tool.poetry]
name = "nonebot-plugin-BotMailNotice"
version = "0.0.1"
description = "基于nonebot2bot断连时的Mail通知插件"
authors = ["ZM25XC <xingling25@qq.com>"]
license="MIT"
readme="README.md"
homepage="https://github.com/ZM25XC/BotMailNotice"
packages = [
  { include = "nonebot-plugin-BotMailNotice" },
]
[tool.poetry.dependencies]
python = "^3.8"
nonebot2="^2.0.0"
aiosmtplib="^2.0.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"